+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Allow disabling props blocks and slots parsing in the `Options` to `comark-mdx`
 - Allow customizing output for labels separate from slots in the `Options` to `comark-mdx`
+- If a props block is an array at the top level, we now produce a warning and discard it instead of silently emitting broken output
+- Fix failing to parse multiple named slots within the same paragraph
 
 ## 0.1.1 (2026-09-12)